makefile not working without libft

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: hthomas <hthomas@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/27 13:44:08 by hthomas           #+#    #+#             */
-/*   Updated: 2019/12/06 14:52:58 by hthomas          ###   ########.fr       */
+/*   Updated: 2019/12/06 15:58:51 by hthomas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,8 +36,8 @@ typedef struct		s_flags
 {
 	int				zb;
 	int				za;
-	int				width;
-	int				pr;
+	unsigned long	width;
+	unsigned long	pr;
 }					t_f;
 
 int					ft_printf(const char *format, ...);
